Introduce Loadable, Cachable, Storable and Hashable for Option<T>
